Inference Flow: updating ext filters for HCM and route.3
