+++ src/backends/envoy_xds_backend/mod.rs
@@ -1,4 +1,3 @@
-mod converters;
 mod envoy_deployer;
 mod model;
 mod resources;
